Blockクラス群の継承関係を決定
blockのkill動作以外の処理が正しく動く
