+++ ts/menu/menu.ts
@@ -1,4 +1,5 @@
-import { Block } from "../block/block.js";
+import { FunBlock } from "../block/fun_block.js";
+import { I32Block } from "../block/i32_block.js";
 import { MenuBlock } from "./menu_block.js";
 
 export class MenuManager {
@@ -34,10 +35,10 @@ export class MenuManager {
     private event_mousedown(_: MouseEvent): void {
         //! [TODO]
         //! debug
-        /*document.getElementById("blocks").appendChild(new FunBlock(window.innerWidth * 0.5, window.innerHeight * 0.5, "+", { args_cnt: 2, return_type: { type_id: PyramidTypeID.I32, attribute: null } }));
-        document.getElementById("blocks").appendChild(new AtomBlock(window.innerWidth * 0.5 + 100, window.innerHeight * 0.5 + 100, "12", PyramidTypeID.I32));
-        document.getElementById("blocks").appendChild(new AtomBlock(window.innerWidth * 0.5 + 200, window.innerHeight * 0.5 + 200, "3", PyramidTypeID.I32));
-        */
+        document.getElementById("blocks").appendChild(new FunBlock(window.innerWidth * 0.5, window.innerHeight * 0.5, "+", { args_cnt: 2, return_type: { type_id: PyramidTypeID.I32, attribute: null } }));
+        document.getElementById("blocks").appendChild(new I32Block(window.innerWidth * 0.5 + 100, window.innerHeight * 0.5 + 100, "12"));
+        document.getElementById("blocks").appendChild(new I32Block(window.innerWidth * 0.5 + 200, window.innerHeight * 0.5 + 200, "3"));
+        //*/
     }       
 
     add_menu_content(menu_content: PyramidMenuContent): void {
